Refactor admin portal modules
