+++ backend/admin/app.js
@@ -1,587 +1,40 @@
+import {
+  createSupabaseClient,
+  getBrowserLocale,
+  getSupabaseLoadMessage,
+} from './supabase-client.js';
+import { createI18n, languageOptions } from './i18n.js';
+import {
+  dayCodeOptions,
+  planStatuses,
+  templateDayOptions,
+  templateSlotsPerDay,
+} from './constants.js';
+
 (() => {
   const { createApp, ref, computed, onMounted, watch } = Vue;
 
-  // === Configure your Supabase project ===
-  const SUPABASE_URL = 'https://jrqjysycoqhlnyufhliy.supabase.co';
-  const SUPABASE_ANON = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImpycWp5c3ljb3FobG55dWZobGl5Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3NTI0MzM0NTIsImV4cCI6MjA2ODAwOTQ1Mn0.3BVA-Ar9YtLGGO12Gt6NQkMl2cn18E_b48PGtlFxxCw';
-  const g = window;
-  const createClient =
-    (g.supabase && g.supabase.createClient) ||
-    (g.Supabase && g.Supabase.createClient) ||
-    g.SUPABASE_CREATE_CLIENT;
-  const fallbackLocale = (navigator.language || '').toLowerCase().startsWith('it')
-    ? 'it'
-    : 'en';
-  const supabaseLoadMessages = {
-    en: 'Supabase library failed to load. Check your network or replace with local copies.',
-    it: 'Impossibile caricare la libreria Supabase. Controlla la rete o sostituiscila con copie locali.',
-  };
-  if (!createClient) {
-    console.error('Supabase JS failed to load from all CDNs (jsDelivr/ESM).');
-    alert(supabaseLoadMessages[fallbackLocale]);
-    return; // stop bootstrapping
+  const fallbackLocale = getBrowserLocale();
+  const { supabase, error } = createSupabaseClient();
+  if (!supabase) {
+    console.error(error);
+    alert(getSupabaseLoadMessage(fallbackLocale));
+    return;
   }
-  const supabase = createClient(SUPABASE_URL, SUPABASE_ANON);
 
   createApp({
     setup() {
       const storedLocale = localStorage.getItem('adminLocale');
-      const browserLocale = (navigator.language || '').toLowerCase().startsWith('it')
-        ? 'it'
-        : 'en';
-      const locale = ref(storedLocale || browserLocale);
-      const languageOptions = [
-        { value: 'en', label: 'English' },
-        { value: 'it', label: 'Italiano' },
-      ];
-      const translations = {
-        en: {
-          app: {
-            title: 'Calisync Admin Portal',
-          },
-          auth: {
-            subtitle:
-              'Sign in with your Supabase account to manage trainees, their workout days, and exercises.',
-          },
-          toolbar: {
-            language: 'Language',
-            admin: 'Admin',
-            roleAdmin: 'Admin',
-            roleTrainer: 'Trainer',
-            roleViewer: 'Viewer',
-            searchPlaceholder: 'Search trainees...',
-          },
-          sections: {
-            dashboard: 'Overview',
-            trainees: 'Trainees',
-            payments: 'Payments',
-            program: 'Program',
-            exercises: 'Exercises',
-            schedule: 'Schedule',
-          },
-          actions: {
-            signIn: 'Sign in',
-            signOut: 'Sign out',
-            save: 'Save',
-            savePlan: 'Save plan',
-            saveDay: 'Save day',
-            reset: 'Reset',
-            delete: 'Delete',
-            add: 'Add',
-            addExercise: 'Add exercise',
-            assignTrainer: 'Assign trainer',
-            refresh: 'Refresh',
-            clear: 'Clear',
-            useNextWeek: 'Use next week',
-            openSchedule: 'Open schedule',
-            loadDays: 'Load days',
-            loadPlans: 'Load plans',
-            collapse: 'Collapse',
-            expand: 'Expand',
-          },
-          labels: {
-            totalCount: '{count} total',
-            shownCount: '{count} shown',
-            dayCount: '{count} day',
-            daysCount: '{count} days',
-            exerciseCount: '{count} exercise',
-            exercisesCount: '{count} exercises',
-            testCount: '{count} test',
-            testsCount: '{count} tests',
-            exerciseShort: 'ex',
-            name: 'Name',
-            status: 'Status',
-            startsAt: 'Starts at',
-            endsAt: 'Ends at',
-            notes: 'Notes',
-            week: 'Week',
-            dayCode: 'Day code',
-            title: 'Title',
-            position: 'Position',
-            weeks: 'Weeks',
-            days: 'Days',
-            exercises: 'Exercises',
-            daysWithExercises: 'Days w/ exercises',
-            weekNumber: 'Week {week}',
-            weekDay: 'Week {week} • {day}',
-            weekDayTitle: 'Week {week} • {day} — {title}',
-            untitled: 'Untitled',
-            exercise: 'Exercise',
-            day: 'Day',
-            unknownExercise: 'Unknown exercise',
-          },
-          dashboard: {
-            feedbackTitle: 'Feedback & trainee notes',
-            feedbackSubtitle: 'Latest messages that need your attention first.',
-            loadingFeedback: 'Loading feedback…',
-            noFeedback: 'No notes yet from trainees.',
-            noticesTitle: 'General notices',
-            noticesSubtitle: 'Quick reminders for upcoming changes or deloads.',
-            noticePlaceholder: 'Add a notice, e.g. “Igor Monday deload week”.',
-            addNotice: 'Add notice',
-            noNotices: 'No general notices yet.',
-            openProgram: 'Open program',
-          },
-          payments: {
-            overviewTitle: 'Payment dashboard',
-            overviewSubtitle: 'Check who is up to date and who needs a reminder.',
-            manageTitle: 'Manage payments',
-            manageSubtitle: 'Update payment status for each trainee.',
-            overdueTitle: 'Overdue reminders',
-            overdueSubtitle: 'Follow up or mark payments as received.',
-            totalLabel: 'Trainees',
-            paidLabel: 'On time',
-            overdueLabel: 'Overdue',
-            noOverdue: 'All payments are up to date.',
-            noMatches: 'No trainees match this filter yet.',
-            filterAll: 'All',
-            filterPaid: 'On time',
-            filterOverdue: 'Overdue',
-            markPaid: 'Mark as paid',
-          },
-          program: {
-            title: 'Program creation',
-            titleWithName: 'Program creation • {name}',
-            subtitle: 'Create workout structures, then map them into days.',
-            empty: 'Select a trainee from the Trainees tab to start building a program.',
-            templateTitle: 'Base template',
-            templateSubtitle:
-              'Sketch the week structure first, then refine with the daily schedule.',
-            dayCountLabel: 'Training days',
-            templateDay: 'Day {day}',
-            templateExercisePlaceholder: 'Exercise',
-            templateSetsPlaceholder: 'Sets/Reps',
-            templateNotesPlaceholder: 'Notes',
-          },
-          placeholders: {
-            email: 'Email',
-            password: 'Password',
-            exerciseExample: 'e.g. Push-ups',
-            exerciseName: 'Exercise name',
-            planExample: 'e.g. Summer Strength',
-            planNotes: 'Optional notes for the trainee',
-            notesOptional: 'Optional notes',
-            notesOptionalShort: 'Notes (optional)',
-            workoutTitle: 'Workout title',
-            filterExercises: 'Type to filter exercises',
-          },
-          exercises: {
-            available: 'Available exercises',
-            none: 'No exercises loaded yet.',
-          },
-          trainees: {
-            badge: 'trainee',
-          },
-          trainers: {
-            title: 'Trainers',
-            assigned: 'Assigned trainers',
-            none: 'No trainers assigned',
-            select: 'Select trainer',
-          },
-          payment: {
-            onTime: 'Payments on time',
-            overdue: 'Payment overdue',
-            toggle: 'On time',
-          },
-          status: {
-            savingExercise: 'Saving exercise…',
-            savingPlan: 'Saving plan…',
-            savingDay: 'Saving day…',
-            refreshingProgress: 'Refreshing progress…',
-            updatingPayment: 'Updating payment status…',
-            updatingTrainer: 'Updating trainer assignment…',
-            noProgress: 'No progress logged yet.',
-            loadingMaxTests: 'Loading max tests…',
-          },
-          plans: {
-            title: 'Workout plans',
-            empty: 'Select a trainee from the Trainees tab to manage their plans.',
-            addTitle: 'Add workout plan',
-            listTitle: 'Plans',
-            none: 'No plans for this trainee yet.',
-          },
-          schedule: {
-            empty: 'Pick a trainee from the Trainees tab to manage their schedule.',
-            createDay: 'Create day',
-            nextWeek: 'Next: week {week}',
-            quickDayPick: 'Quick day pick',
-            quickDayHelp: 'Tap a code to fill the day quickly.',
-            recap: 'Schedule recap',
-            recapSubtitle: 'Quick snapshot of what has been added.',
-            recapEmpty: 'Add a day to start building the recap.',
-            highlights: 'Highlights',
-            jumpToDay: 'Jump to day',
-            jumpSubtitle: 'Move fast between days and auto-expand the one you need.',
-            noDays: 'No days yet.',
-            daysExercises: 'Days & Exercises',
-            addExercise: 'Add exercise',
-            searchSelect: 'Search & select',
-            exerciseHelp: 'Start typing to auto-complete and click a suggestion.',
-          },
-          history: {
-            title: 'Max test history',
-            empty: 'Select a trainee from the Trainees tab to review their max test history.',
-            titleWithName: 'Max test history • {name}',
-            subtitle:
-              'Each chart shows how max attempts evolve over time for a single exercise.',
-            none: 'No max tests logged yet.',
-            testsBest: '{countLabel} • best {value} {unit}',
-            chartAria: 'Max test history for {exercise}',
-          },
-          errors: {
-            loadTrainees: 'Failed to load trainees: {message}',
-            updatePayment: 'Failed to update payment status.',
-            loadAccess: 'Failed to load admin access: {message}',
-            loadTrainers: 'Failed to load trainers: {message}',
-            assignTrainer: 'Failed to assign trainer.',
-            removeTrainer: 'Failed to remove trainer.',
-            loadExercises: 'Failed to load exercises: {message}',
-            loadProgress: 'Failed to load trainee progress: {message}',
-            loadPlans: 'Failed to load plans: {message}',
-            exerciseNameRequired: 'Exercise name is required.',
-            exerciseNameEmpty: 'Exercise name cannot be empty.',
-            createExercise: 'Failed to create exercise.',
-            updateExercise: 'Failed to update exercise.',
-            deleteExercise: 'Failed to delete exercise.',
-            loadDays: 'Failed to load days: {message}',
-            selectTrainee: 'Select a trainee first.',
-            planNameRequired: 'Plan name is required.',
-            createPlan: 'Failed to create plan.',
-            dayCodeRequired: 'Day code is required.',
-            createDay: 'Failed to create day.',
-            missingPlan: 'Missing plan.',
-            updatePlan: 'Failed to update plan.',
-            deletePlan: 'Failed to delete plan.',
-            missingDay: 'Missing day.',
-            chooseExercise: 'Choose an exercise first.',
-            addExercise: 'Failed to add exercise.',
-            updateDay: 'Failed to update day: {message}',
-            deleteDay: 'Failed to delete day: {message}',
-            missingDayExercise: 'Missing day exercise.',
-            updateDayExercise: 'Failed to update exercise: {message}',
-            deleteDayExercise: 'Failed to delete exercise: {message}',
-            loadMaxTests: 'Failed to load max tests.',
-            loadMaxTestsWithMessage: 'Failed to load max tests: {message}',
-          },
-          confirm: {
-            deleteExercise: 'Delete exercise "{name}"?',
-            deletePlan: 'Delete this plan?',
-            deleteDay: 'Delete this day and its exercises?',
-            deleteDayExercise: 'Remove exercise from day?',
-          },
-          dayCodes: {
-            MON: 'Mon',
-            TUE: 'Tue',
-            WED: 'Wed',
-            THU: 'Thu',
-            FRI: 'Fri',
-            SAT: 'Sat',
-            SUN: 'Sun',
-          },
-          planStatuses: {
-            active: 'active',
-            upcoming: 'upcoming',
-            draft: 'draft',
-            archived: 'archived',
-          },
-        },
-        it: {
-          app: {
-            title: 'Portale Admin Calisync',
-          },
-          auth: {
-            subtitle:
-              'Accedi con il tuo account Supabase per gestire gli allievi, le loro giornate di allenamento e gli esercizi.',
-          },
-          toolbar: {
-            language: 'Lingua',
-            admin: 'Admin',
-            roleAdmin: 'Admin',
-            roleTrainer: 'Trainer',
-            roleViewer: 'Visualizzatore',
-            searchPlaceholder: 'Cerca allievi...',
-          },
-          sections: {
-            dashboard: 'Panoramica',
-            trainees: 'Allievi',
-            payments: 'Pagamenti',
-            program: 'Programma',
-            exercises: 'Esercizi',
-            schedule: 'Programma',
-          },
-          actions: {
-            signIn: 'Accedi',
-            signOut: 'Esci',
-            save: 'Salva',
-            savePlan: 'Salva piano',
-            saveDay: 'Salva giornata',
-            reset: 'Reimposta',
-            delete: 'Elimina',
-            add: 'Aggiungi',
-            addExercise: 'Aggiungi esercizio',
-            assignTrainer: 'Assegna trainer',
-            refresh: 'Aggiorna',
-            clear: 'Svuota',
-            useNextWeek: 'Usa prossima settimana',
-            openSchedule: 'Apri programma',
-            loadDays: 'Carica giornate',
-            loadPlans: 'Carica piani',
-            collapse: 'Comprimi',
-            expand: 'Espandi',
-          },
-          labels: {
-            totalCount: '{count} totali',
-            shownCount: '{count} mostrati',
-            dayCount: '{count} giorno',
-            daysCount: '{count} giorni',
-            exerciseCount: '{count} esercizio',
-            exercisesCount: '{count} esercizi',
-            testCount: '{count} test',
-            testsCount: '{count} test',
-            exerciseShort: 'es',
-            name: 'Nome',
-            status: 'Stato',
-            startsAt: 'Inizio',
-            endsAt: 'Fine',
-            notes: 'Note',
-            week: 'Settimana',
-            dayCode: 'Codice giorno',
-            title: 'Titolo',
-            position: 'Posizione',
-            weeks: 'Settimane',
-            days: 'Giorni',
-            exercises: 'Esercizi',
-            daysWithExercises: 'Giorni con esercizi',
-            weekNumber: 'Settimana {week}',
-            weekDay: 'Settimana {week} • {day}',
-            weekDayTitle: 'Settimana {week} • {day} — {title}',
-            untitled: 'Senza titolo',
-            exercise: 'Esercizio',
-            day: 'Giorno',
-            unknownExercise: 'Esercizio sconosciuto',
-          },
-          dashboard: {
-            feedbackTitle: 'Feedback & note allievi',
-            feedbackSubtitle: 'Messaggi recenti da vedere subito.',
-            loadingFeedback: 'Caricamento feedback…',
-            noFeedback: 'Nessuna nota inviata dagli allievi.',
-            noticesTitle: 'Avvisi generali',
-            noticesSubtitle: 'Promemoria veloci per scarichi o cambiamenti.',
-            noticePlaceholder: 'Aggiungi un avviso, es. “Igor lunedì settimana di scarico”.',
-            addNotice: 'Aggiungi avviso',
-            noNotices: 'Nessun avviso generale.',
-            openProgram: 'Apri programma',
-          },
-          payments: {
-            overviewTitle: 'Dashboard pagamenti',
-            overviewSubtitle: 'Controlla chi è in regola e chi necessita un promemoria.',
-            manageTitle: 'Gestione pagamenti',
-            manageSubtitle: 'Aggiorna lo stato di pagamento degli allievi.',
-            overdueTitle: 'Promemoria in ritardo',
-            overdueSubtitle: 'Contatta o segna i pagamenti ricevuti.',
-            totalLabel: 'Allievi',
-            paidLabel: 'In regola',
-            overdueLabel: 'In ritardo',
-            noOverdue: 'Tutti i pagamenti sono in regola.',
-            noMatches: 'Nessun allievo corrisponde a questo filtro.',
-            filterAll: 'Tutti',
-            filterPaid: 'In regola',
-            filterOverdue: 'In ritardo',
-            markPaid: 'Segna pagato',
-          },
-          program: {
-            title: 'Creazione programma',
-            titleWithName: 'Creazione programma • {name}',
-            subtitle: 'Crea la struttura base e poi rifinisci le giornate.',
-            empty: 'Seleziona un allievo dalla scheda Allievi per creare il programma.',
-            templateTitle: 'Scheda base',
-            templateSubtitle:
-              'Imposta prima la struttura, poi rifinisci con le giornate.',
-            dayCountLabel: 'Giorni di allenamento',
-            templateDay: 'Giorno {day}',
-            templateExercisePlaceholder: 'Esercizio',
-            templateSetsPlaceholder: 'Serie/Rip.',
-            templateNotesPlaceholder: 'Note',
-          },
-          placeholders: {
-            email: 'Email',
-            password: 'Password',
-            exerciseExample: 'es. Push-up',
-            exerciseName: 'Nome esercizio',
-            planExample: 'es. Forza estiva',
-            planNotes: "Note opzionali per l'allievo",
-            notesOptional: 'Note opzionali',
-            notesOptionalShort: 'Note (opzionali)',
-            workoutTitle: 'Titolo allenamento',
-            filterExercises: 'Digita per filtrare esercizi',
-          },
-          exercises: {
-            available: 'Esercizi disponibili',
-            none: 'Nessun esercizio caricato.',
-          },
-          trainees: {
-            badge: 'allievo',
-          },
-          trainers: {
-            title: 'Trainer',
-            assigned: 'Trainer assegnati',
-            none: 'Nessun trainer assegnato',
-            select: 'Seleziona trainer',
-          },
-          payment: {
-            onTime: 'Pagamenti regolari',
-            overdue: 'Pagamento in ritardo',
-            toggle: 'In regola',
-          },
-          status: {
-            savingExercise: 'Salvataggio esercizio…',
-            savingPlan: 'Salvataggio piano…',
-            savingDay: 'Salvataggio giornata…',
-            refreshingProgress: 'Aggiornamento progressi…',
-            updatingPayment: 'Aggiornamento stato pagamento…',
-            updatingTrainer: 'Aggiornamento assegnazione trainer…',
-            noProgress: 'Nessun progresso registrato.',
-            loadingMaxTests: 'Caricamento test massimali…',
-          },
-          plans: {
-            title: 'Piani di allenamento',
-            empty: 'Seleziona un allievo dalla scheda Allievi per gestire i piani.',
-            addTitle: 'Aggiungi piano di allenamento',
-            listTitle: 'Piani',
-            none: 'Nessun piano per questo allievo.',
-          },
-          schedule: {
-            empty: 'Seleziona un allievo dalla scheda Allievi per gestire il programma.',
-            createDay: 'Crea giornata',
-            nextWeek: 'Prossima: settimana {week}',
-            quickDayPick: 'Scelta rapida giornata',
-            quickDayHelp: 'Tocca un codice per compilare rapidamente la giornata.',
-            recap: 'Riepilogo programma',
-            recapSubtitle: 'Vista rapida di ciò che è stato aggiunto.',
-            recapEmpty: 'Aggiungi una giornata per iniziare il riepilogo.',
-            highlights: 'In evidenza',
-            jumpToDay: 'Vai alla giornata',
-            jumpSubtitle:
-              'Passa rapidamente tra le giornate ed espandi quella che ti serve.',
-            noDays: 'Nessuna giornata ancora.',
-            daysExercises: 'Giornate ed esercizi',
-            addExercise: 'Aggiungi esercizio',
-            searchSelect: 'Cerca e seleziona',
-            exerciseHelp:
-              'Inizia a digitare per completare automaticamente e fai clic su un suggerimento.',
-          },
-          history: {
-            title: 'Storico test massimali',
-            empty:
-              'Seleziona un allievo dalla scheda Allievi per vedere lo storico dei test massimali.',
-            titleWithName: 'Storico test massimali • {name}',
-            subtitle:
-              'Ogni grafico mostra come evolvono i massimali nel tempo per un singolo esercizio.',
-            none: 'Nessun test massimale registrato.',
-            testsBest: '{countLabel} • migliore {value} {unit}',
-            chartAria: 'Storico test massimali per {exercise}',
-          },
-          errors: {
-            loadTrainees: 'Impossibile caricare gli allievi: {message}',
-            updatePayment: 'Impossibile aggiornare lo stato pagamento.',
-            loadAccess: 'Impossibile caricare i permessi admin: {message}',
-            loadTrainers: 'Impossibile caricare i trainer: {message}',
-            assignTrainer: 'Impossibile assegnare il trainer.',
-            removeTrainer: 'Impossibile rimuovere il trainer.',
-            loadExercises: 'Impossibile caricare gli esercizi: {message}',
-            loadProgress: 'Impossibile caricare i progressi degli allievi: {message}',
-            loadPlans: 'Impossibile caricare i piani: {message}',
-            exerciseNameRequired: "Il nome dell'esercizio è obbligatorio.",
-            exerciseNameEmpty: "Il nome dell'esercizio non può essere vuoto.",
-            createExercise: "Impossibile creare l'esercizio.",
-            updateExercise: "Impossibile aggiornare l'esercizio.",
-            deleteExercise: "Impossibile eliminare l'esercizio.",
-            loadDays: 'Impossibile caricare le giornate: {message}',
-            selectTrainee: 'Seleziona prima un allievo.',
-            planNameRequired: 'Il nome del piano è obbligatorio.',
-            createPlan: 'Impossibile creare il piano.',
-            dayCodeRequired: 'Il codice giorno è obbligatorio.',
-            createDay: 'Impossibile creare la giornata.',
-            missingPlan: 'Piano mancante.',
-            updatePlan: 'Impossibile aggiornare il piano.',
-            deletePlan: 'Impossibile eliminare il piano.',
-            missingDay: 'Giornata mancante.',
-            chooseExercise: 'Seleziona prima un esercizio.',
-            addExercise: "Impossibile aggiungere l'esercizio.",
-            updateDay: 'Impossibile aggiornare la giornata: {message}',
-            deleteDay: 'Impossibile eliminare la giornata: {message}',
-            missingDayExercise: 'Esercizio del giorno mancante.',
-            updateDayExercise: "Impossibile aggiornare l'esercizio: {message}",
-            deleteDayExercise: "Impossibile eliminare l'esercizio: {message}",
-            loadMaxTests: 'Impossibile caricare i test massimali.',
-            loadMaxTestsWithMessage:
-              'Impossibile caricare i test massimali: {message}',
-          },
-          confirm: {
-            deleteExercise: 'Eliminare l’esercizio "{name}"?',
-            deletePlan: 'Eliminare questo piano?',
-            deleteDay: 'Eliminare questa giornata e i suoi esercizi?',
-            deleteDayExercise: "Rimuovere l'esercizio dalla giornata?",
-          },
-          dayCodes: {
-            MON: 'Lun',
-            TUE: 'Mar',
-            WED: 'Mer',
-            THU: 'Gio',
-            FRI: 'Ven',
-            SAT: 'Sab',
-            SUN: 'Dom',
-          },
-          planStatuses: {
-            active: 'attivo',
-            upcoming: 'in arrivo',
-            draft: 'bozza',
-            archived: 'archiviato',
-          },
-        },
-      };
-      const interpolate = (template, params) =>
-        template.replace(/\{(\w+)\}/g, (_match, key) =>
-          params[key] !== undefined ? params[key] : '',
-        );
-      const getTranslation = (key, selectedLocale) => {
-        const segments = key.split('.');
-        let current = translations[selectedLocale];
-        for (const segment of segments) {
-          if (!current || typeof current !== 'object' || !(segment in current)) {
-            return null;
-          }
-          current = current[segment];
-        }
-        return typeof current === 'string' ? current : null;
-      };
-      const t = (key, params = {}) => {
-        const selectedLocale = locale.value in translations ? locale.value : 'en';
-        const translation =
-          getTranslation(key, selectedLocale) || getTranslation(key, 'en') || key;
-        return interpolate(translation, params);
-      };
-      const formatCount = (count, singularKey, pluralKey) =>
-        t(count === 1 ? singularKey : pluralKey, { count });
-      const dayCodeLabel = (code) => {
-        const normalized = (code || '').toUpperCase();
-        const label =
-          translations[locale.value]?.dayCodes?.[normalized] ||
-          translations.en?.dayCodes?.[normalized];
-        return label || normalized || t('labels.day');
-      };
-      const planStatusLabel = (status) =>
-        t(`planStatuses.${status}`, { status });
-      const formatWeekDayLabel = (week, code) =>
-        t('labels.weekDay', { week, day: dayCodeLabel(code) });
-      const formatWeekDayTitleLabel = (week, code, title) =>
-        title
-          ? t('labels.weekDayTitle', { week, day: dayCodeLabel(code), title })
-          : formatWeekDayLabel(week, code);
-      const updateDocumentLanguage = () => {
-        document.documentElement.lang = locale.value;
-        document.title = t('app.title');
-      };
+      const locale = ref(storedLocale || fallbackLocale);
+      const {
+        t,
+        formatCount,
+        dayCodeLabel,
+        planStatusLabel,
+        formatWeekDayLabel,
+        formatWeekDayTitleLabel,
+        updateDocumentLanguage,
+      } = createI18n(locale);
 
       watch(locale, (nextLocale) => {
         localStorage.setItem('adminLocale', nextLocale);
@@ -631,16 +84,12 @@
       const newDayTitle = ref('');
       const newDayNotes = ref('');
       const newExerciseName = ref('');
-      const planStatuses = ['active', 'upcoming', 'draft', 'archived'];
-      const dayCodeOptions = ['MON', 'TUE', 'WED', 'THU', 'FRI', 'SAT', 'SUN'];
       const newPlanName = ref('');
       const newPlanStatus = ref(planStatuses[0]);
       const newPlanStartsAt = ref('');
       const newPlanEndsAt = ref('');
       const newPlanNotes = ref('');
       const templateDayCount = ref(3);
-      const templateDayOptions = [2, 3, 4, 5];
-      const templateSlotsPerDay = 6;
       const programTemplateDays = ref(
         buildTemplateDays(templateDayCount.value, templateSlotsPerDay, []),
       );
