+++ backend/admin/constants.js
@@ -1,0 +1,4 @@
+export const planStatuses = ['active', 'upcoming', 'draft', 'archived'];
+export const dayCodeOptions = ['MON', 'TUE', 'WED', 'THU', 'FRI', 'SAT', 'SUN'];
+export const templateDayOptions = [2, 3, 4, 5];
+export const templateSlotsPerDay = 6;
