+++ backend/admin/index.html
@@ -661,6 +661,6 @@
     })();
 </script>
 <script type="module" src="./supabase-bridge.js"></script>
-<script src="./app.js"></script>
+<script type="module" src="./app.js"></script>
 </body>
 </html>
