+++ backend/admin/translations.js
@@ -1,0 +1,502 @@
+export const translations = {
+  en: {
+    app: {
+      title: 'Calisync Admin Portal',
+    },
+    auth: {
+      subtitle:
+        'Sign in with your Supabase account to manage trainees, their workout days, and exercises.',
+    },
+    toolbar: {
+      language: 'Language',
+      admin: 'Admin',
+      roleAdmin: 'Admin',
+      roleTrainer: 'Trainer',
+      roleViewer: 'Viewer',
+      searchPlaceholder: 'Search trainees...',
+    },
+    sections: {
+      dashboard: 'Overview',
+      trainees: 'Trainees',
+      payments: 'Payments',
+      program: 'Program',
+      exercises: 'Exercises',
+      schedule: 'Schedule',
+    },
+    actions: {
+      signIn: 'Sign in',
+      signOut: 'Sign out',
+      save: 'Save',
+      savePlan: 'Save plan',
+      saveDay: 'Save day',
+      reset: 'Reset',
+      delete: 'Delete',
+      add: 'Add',
+      addExercise: 'Add exercise',
+      assignTrainer: 'Assign trainer',
+      refresh: 'Refresh',
+      clear: 'Clear',
+      useNextWeek: 'Use next week',
+      openSchedule: 'Open schedule',
+      loadDays: 'Load days',
+      loadPlans: 'Load plans',
+      collapse: 'Collapse',
+      expand: 'Expand',
+    },
+    labels: {
+      totalCount: '{count} total',
+      shownCount: '{count} shown',
+      dayCount: '{count} day',
+      daysCount: '{count} days',
+      exerciseCount: '{count} exercise',
+      exercisesCount: '{count} exercises',
+      testCount: '{count} test',
+      testsCount: '{count} tests',
+      exerciseShort: 'ex',
+      name: 'Name',
+      status: 'Status',
+      startsAt: 'Starts at',
+      endsAt: 'Ends at',
+      notes: 'Notes',
+      week: 'Week',
+      dayCode: 'Day code',
+      title: 'Title',
+      position: 'Position',
+      weeks: 'Weeks',
+      days: 'Days',
+      exercises: 'Exercises',
+      daysWithExercises: 'Days w/ exercises',
+      weekNumber: 'Week {week}',
+      weekDay: 'Week {week} • {day}',
+      weekDayTitle: 'Week {week} • {day} — {title}',
+      untitled: 'Untitled',
+      exercise: 'Exercise',
+      day: 'Day',
+      unknownExercise: 'Unknown exercise',
+    },
+    dashboard: {
+      feedbackTitle: 'Feedback & trainee notes',
+      feedbackSubtitle: 'Latest messages that need your attention first.',
+      loadingFeedback: 'Loading feedback…',
+      noFeedback: 'No notes yet from trainees.',
+      noticesTitle: 'General notices',
+      noticesSubtitle: 'Quick reminders for upcoming changes or deloads.',
+      noticePlaceholder: 'Add a notice, e.g. “Igor Monday deload week”.',
+      addNotice: 'Add notice',
+      noNotices: 'No general notices yet.',
+      openProgram: 'Open program',
+    },
+    payments: {
+      overviewTitle: 'Payment dashboard',
+      overviewSubtitle: 'Check who is up to date and who needs a reminder.',
+      manageTitle: 'Manage payments',
+      manageSubtitle: 'Update payment status for each trainee.',
+      overdueTitle: 'Overdue reminders',
+      overdueSubtitle: 'Follow up or mark payments as received.',
+      totalLabel: 'Trainees',
+      paidLabel: 'On time',
+      overdueLabel: 'Overdue',
+      noOverdue: 'All payments are up to date.',
+      noMatches: 'No trainees match this filter yet.',
+      filterAll: 'All',
+      filterPaid: 'On time',
+      filterOverdue: 'Overdue',
+      markPaid: 'Mark as paid',
+    },
+    program: {
+      title: 'Program creation',
+      titleWithName: 'Program creation • {name}',
+      subtitle: 'Create workout structures, then map them into days.',
+      empty: 'Select a trainee from the Trainees tab to start building a program.',
+      templateTitle: 'Base template',
+      templateSubtitle:
+        'Sketch the week structure first, then refine with the daily schedule.',
+      dayCountLabel: 'Training days',
+      templateDay: 'Day {day}',
+      templateExercisePlaceholder: 'Exercise',
+      templateSetsPlaceholder: 'Sets/Reps',
+      templateNotesPlaceholder: 'Notes',
+    },
+    placeholders: {
+      email: 'Email',
+      password: 'Password',
+      exerciseExample: 'e.g. Push-ups',
+      exerciseName: 'Exercise name',
+      planExample: 'e.g. Summer Strength',
+      planNotes: 'Optional notes for the trainee',
+      notesOptional: 'Optional notes',
+      notesOptionalShort: 'Notes (optional)',
+      workoutTitle: 'Workout title',
+      filterExercises: 'Type to filter exercises',
+    },
+    exercises: {
+      available: 'Available exercises',
+      none: 'No exercises loaded yet.',
+    },
+    trainees: {
+      badge: 'trainee',
+    },
+    trainers: {
+      title: 'Trainers',
+      assigned: 'Assigned trainers',
+      none: 'No trainers assigned',
+      select: 'Select trainer',
+    },
+    payment: {
+      onTime: 'Payments on time',
+      overdue: 'Payment overdue',
+      toggle: 'On time',
+    },
+    status: {
+      savingExercise: 'Saving exercise…',
+      savingPlan: 'Saving plan…',
+      savingDay: 'Saving day…',
+      refreshingProgress: 'Refreshing progress…',
+      updatingPayment: 'Updating payment status…',
+      updatingTrainer: 'Updating trainer assignment…',
+      noProgress: 'No progress logged yet.',
+      loadingMaxTests: 'Loading max tests…',
+    },
+    plans: {
+      title: 'Workout plans',
+      empty: 'Select a trainee from the Trainees tab to manage their plans.',
+      addTitle: 'Add workout plan',
+      listTitle: 'Plans',
+      none: 'No plans for this trainee yet.',
+    },
+    schedule: {
+      empty: 'Pick a trainee from the Trainees tab to manage their schedule.',
+      createDay: 'Create day',
+      nextWeek: 'Next: week {week}',
+      quickDayPick: 'Quick day pick',
+      quickDayHelp: 'Tap a code to fill the day quickly.',
+      recap: 'Schedule recap',
+      recapSubtitle: 'Quick snapshot of what has been added.',
+      recapEmpty: 'Add a day to start building the recap.',
+      highlights: 'Highlights',
+      jumpToDay: 'Jump to day',
+      jumpSubtitle: 'Move fast between days and auto-expand the one you need.',
+      noDays: 'No days yet.',
+      daysExercises: 'Days & Exercises',
+      addExercise: 'Add exercise',
+      searchSelect: 'Search & select',
+      exerciseHelp: 'Start typing to auto-complete and click a suggestion.',
+    },
+    history: {
+      title: 'Max test history',
+      empty: 'Select a trainee from the Trainees tab to review their max test history.',
+      titleWithName: 'Max test history • {name}',
+      subtitle:
+        'Each chart shows how max attempts evolve over time for a single exercise.',
+      none: 'No max tests logged yet.',
+      testsBest: '{countLabel} • best {value} {unit}',
+      chartAria: 'Max test history for {exercise}',
+    },
+    errors: {
+      loadTrainees: 'Failed to load trainees: {message}',
+      updatePayment: 'Failed to update payment status.',
+      loadAccess: 'Failed to load admin access: {message}',
+      loadTrainers: 'Failed to load trainers: {message}',
+      assignTrainer: 'Failed to assign trainer.',
+      removeTrainer: 'Failed to remove trainer.',
+      loadExercises: 'Failed to load exercises: {message}',
+      loadProgress: 'Failed to load trainee progress: {message}',
+      loadPlans: 'Failed to load plans: {message}',
+      exerciseNameRequired: 'Exercise name is required.',
+      exerciseNameEmpty: 'Exercise name cannot be empty.',
+      createExercise: 'Failed to create exercise.',
+      updateExercise: 'Failed to update exercise.',
+      deleteExercise: 'Failed to delete exercise.',
+      loadDays: 'Failed to load days: {message}',
+      selectTrainee: 'Select a trainee first.',
+      planNameRequired: 'Plan name is required.',
+      createPlan: 'Failed to create plan.',
+      dayCodeRequired: 'Day code is required.',
+      createDay: 'Failed to create day.',
+      missingPlan: 'Missing plan.',
+      updatePlan: 'Failed to update plan.',
+      deletePlan: 'Failed to delete plan.',
+      missingDay: 'Missing day.',
+      chooseExercise: 'Choose an exercise first.',
+      addExercise: 'Failed to add exercise.',
+      updateDay: 'Failed to update day: {message}',
+      deleteDay: 'Failed to delete day: {message}',
+      missingDayExercise: 'Missing day exercise.',
+      updateDayExercise: 'Failed to update exercise: {message}',
+      deleteDayExercise: 'Failed to delete exercise: {message}',
+      loadMaxTests: 'Failed to load max tests.',
+      loadMaxTestsWithMessage: 'Failed to load max tests: {message}',
+    },
+    confirm: {
+      deleteExercise: 'Delete exercise "{name}"?',
+      deletePlan: 'Delete this plan?',
+      deleteDay: 'Delete this day and its exercises?',
+      deleteDayExercise: 'Remove exercise from day?',
+    },
+    dayCodes: {
+      MON: 'Mon',
+      TUE: 'Tue',
+      WED: 'Wed',
+      THU: 'Thu',
+      FRI: 'Fri',
+      SAT: 'Sat',
+      SUN: 'Sun',
+    },
+    planStatuses: {
+      active: 'active',
+      upcoming: 'upcoming',
+      draft: 'draft',
+      archived: 'archived',
+    },
+  },
+  it: {
+    app: {
+      title: 'Portale Admin Calisync',
+    },
+    auth: {
+      subtitle:
+        'Accedi con il tuo account Supabase per gestire gli allievi, le loro giornate di allenamento e gli esercizi.',
+    },
+    toolbar: {
+      language: 'Lingua',
+      admin: 'Admin',
+      roleAdmin: 'Admin',
+      roleTrainer: 'Trainer',
+      roleViewer: 'Visualizzatore',
+      searchPlaceholder: 'Cerca allievi...',
+    },
+    sections: {
+      dashboard: 'Panoramica',
+      trainees: 'Allievi',
+      payments: 'Pagamenti',
+      program: 'Programma',
+      exercises: 'Esercizi',
+      schedule: 'Programma',
+    },
+    actions: {
+      signIn: 'Accedi',
+      signOut: 'Esci',
+      save: 'Salva',
+      savePlan: 'Salva piano',
+      saveDay: 'Salva giornata',
+      reset: 'Reimposta',
+      delete: 'Elimina',
+      add: 'Aggiungi',
+      addExercise: 'Aggiungi esercizio',
+      assignTrainer: 'Assegna trainer',
+      refresh: 'Aggiorna',
+      clear: 'Svuota',
+      useNextWeek: 'Usa prossima settimana',
+      openSchedule: 'Apri programma',
+      loadDays: 'Carica giornate',
+      loadPlans: 'Carica piani',
+      collapse: 'Comprimi',
+      expand: 'Espandi',
+    },
+    labels: {
+      totalCount: '{count} totali',
+      shownCount: '{count} mostrati',
+      dayCount: '{count} giorno',
+      daysCount: '{count} giorni',
+      exerciseCount: '{count} esercizio',
+      exercisesCount: '{count} esercizi',
+      testCount: '{count} test',
+      testsCount: '{count} test',
+      exerciseShort: 'es',
+      name: 'Nome',
+      status: 'Stato',
+      startsAt: 'Inizio',
+      endsAt: 'Fine',
+      notes: 'Note',
+      week: 'Settimana',
+      dayCode: 'Codice giorno',
+      title: 'Titolo',
+      position: 'Posizione',
+      weeks: 'Settimane',
+      days: 'Giorni',
+      exercises: 'Esercizi',
+      daysWithExercises: 'Giorni con esercizi',
+      weekNumber: 'Settimana {week}',
+      weekDay: 'Settimana {week} • {day}',
+      weekDayTitle: 'Settimana {week} • {day} — {title}',
+      untitled: 'Senza titolo',
+      exercise: 'Esercizio',
+      day: 'Giorno',
+      unknownExercise: 'Esercizio sconosciuto',
+    },
+    dashboard: {
+      feedbackTitle: 'Feedback & note allievi',
+      feedbackSubtitle: 'Messaggi recenti da vedere subito.',
+      loadingFeedback: 'Caricamento feedback…',
+      noFeedback: 'Nessuna nota inviata dagli allievi.',
+      noticesTitle: 'Avvisi generali',
+      noticesSubtitle: 'Promemoria veloci per scarichi o cambiamenti.',
+      noticePlaceholder: 'Aggiungi un avviso, es. “Igor lunedì settimana di scarico”.',
+      addNotice: 'Aggiungi avviso',
+      noNotices: 'Nessun avviso generale.',
+      openProgram: 'Apri programma',
+    },
+    payments: {
+      overviewTitle: 'Dashboard pagamenti',
+      overviewSubtitle: 'Controlla chi è in regola e chi necessita un promemoria.',
+      manageTitle: 'Gestione pagamenti',
+      manageSubtitle: 'Aggiorna lo stato di pagamento degli allievi.',
+      overdueTitle: 'Promemoria in ritardo',
+      overdueSubtitle: 'Contatta o segna i pagamenti ricevuti.',
+      totalLabel: 'Allievi',
+      paidLabel: 'In regola',
+      overdueLabel: 'In ritardo',
+      noOverdue: 'Tutti i pagamenti sono in regola.',
+      noMatches: 'Nessun allievo corrisponde a questo filtro.',
+      filterAll: 'Tutti',
+      filterPaid: 'In regola',
+      filterOverdue: 'In ritardo',
+      markPaid: 'Segna pagato',
+    },
+    program: {
+      title: 'Creazione programma',
+      titleWithName: 'Creazione programma • {name}',
+      subtitle: 'Crea la struttura base e poi rifinisci le giornate.',
+      empty: 'Seleziona un allievo dalla scheda Allievi per creare il programma.',
+      templateTitle: 'Scheda base',
+      templateSubtitle:
+        'Imposta prima la struttura, poi rifinisci con le giornate.',
+      dayCountLabel: 'Giorni di allenamento',
+      templateDay: 'Giorno {day}',
+      templateExercisePlaceholder: 'Esercizio',
+      templateSetsPlaceholder: 'Serie/Ripetizioni',
+      templateNotesPlaceholder: 'Note',
+    },
+    placeholders: {
+      email: 'Email',
+      password: 'Password',
+      exerciseExample: 'es. Push-up',
+      exerciseName: 'Nome esercizio',
+      planExample: 'es. Forza Estiva',
+      planNotes: 'Note opzionali per l’allievo',
+      notesOptional: 'Note opzionali',
+      notesOptionalShort: 'Note (opzionali)',
+      workoutTitle: 'Titolo workout',
+      filterExercises: 'Digita per filtrare gli esercizi',
+    },
+    exercises: {
+      available: 'Esercizi disponibili',
+      none: 'Nessun esercizio caricato.',
+    },
+    trainees: {
+      badge: 'allievo',
+    },
+    trainers: {
+      title: 'Trainer',
+      assigned: 'Trainer assegnati',
+      none: 'Nessun trainer assegnato',
+      select: 'Seleziona trainer',
+    },
+    payment: {
+      onTime: 'Pagamenti in regola',
+      overdue: 'Pagamento in ritardo',
+      toggle: 'In regola',
+    },
+    status: {
+      savingExercise: 'Salvataggio esercizio…',
+      savingPlan: 'Salvataggio piano…',
+      savingDay: 'Salvataggio giornata…',
+      refreshingProgress: 'Aggiornamento progressi…',
+      updatingPayment: 'Aggiornamento pagamento…',
+      updatingTrainer: 'Aggiornamento assegnazione trainer…',
+      noProgress: 'Nessun progresso registrato.',
+      loadingMaxTests: 'Caricamento test massimali…',
+    },
+    plans: {
+      title: 'Piani allenamento',
+      empty: 'Seleziona un allievo dalla scheda Allievi per gestire i piani.',
+      addTitle: 'Aggiungi piano allenamento',
+      listTitle: 'Piani',
+      none: 'Nessun piano per questo allievo.',
+    },
+    schedule: {
+      empty: 'Seleziona un allievo dalla scheda Allievi per gestire la programmazione.',
+      createDay: 'Crea giornata',
+      nextWeek: 'Prossima: settimana {week}',
+      quickDayPick: 'Selezione rapida',
+      quickDayHelp: 'Tocca un codice per compilare velocemente la giornata.',
+      recap: 'Riepilogo programma',
+      recapSubtitle: 'Panoramica veloce di quanto creato.',
+      recapEmpty: 'Aggiungi una giornata per creare il riepilogo.',
+      highlights: 'Highlights',
+      jumpToDay: 'Vai alla giornata',
+      jumpSubtitle: 'Spostati rapidamente e auto-espandi la giornata necessaria.',
+      noDays: 'Nessuna giornata.',
+      daysExercises: 'Giornate & Esercizi',
+      addExercise: 'Aggiungi esercizio',
+      searchSelect: 'Cerca e seleziona',
+      exerciseHelp: 'Inizia a digitare per completare automaticamente e seleziona un suggerimento.',
+    },
+    history: {
+      title: 'Storico test massimali',
+      empty: 'Seleziona un allievo dalla scheda Allievi per vedere lo storico dei massimali.',
+      titleWithName: 'Storico test massimali • {name}',
+      subtitle:
+        'Ogni grafico mostra l’evoluzione dei massimali per ogni esercizio.',
+      none: 'Nessun test massimale registrato.',
+      testsBest: '{countLabel} • best {value} {unit}',
+      chartAria: 'Storico test massimali per {exercise}',
+    },
+    errors: {
+      loadTrainees: 'Errore caricamento allievi: {message}',
+      updatePayment: 'Impossibile aggiornare il pagamento.',
+      loadAccess: 'Errore caricamento accesso admin: {message}',
+      loadTrainers: 'Errore caricamento trainer: {message}',
+      assignTrainer: 'Impossibile assegnare il trainer.',
+      removeTrainer: 'Impossibile rimuovere il trainer.',
+      loadExercises: 'Errore caricamento esercizi: {message}',
+      loadProgress: 'Errore caricamento progressi: {message}',
+      loadPlans: 'Errore caricamento piani: {message}',
+      exerciseNameRequired: "Il nome dell'esercizio è obbligatorio.",
+      exerciseNameEmpty: "Il nome dell'esercizio non può essere vuoto.",
+      createExercise: 'Impossibile creare esercizio.',
+      updateExercise: 'Impossibile aggiornare esercizio.',
+      deleteExercise: 'Impossibile eliminare esercizio.',
+      loadDays: 'Errore caricamento giornate: {message}',
+      selectTrainee: "Seleziona prima un allievo.",
+      planNameRequired: 'Il nome del piano è obbligatorio.',
+      createPlan: 'Impossibile creare piano.',
+      dayCodeRequired: 'Il codice giorno è obbligatorio.',
+      createDay: 'Impossibile creare giornata.',
+      missingPlan: 'Piano mancante.',
+      updatePlan: 'Impossibile aggiornare piano.',
+      deletePlan: 'Impossibile eliminare piano.',
+      missingDay: 'Giornata mancante.',
+      chooseExercise: "Scegli prima un esercizio.",
+      addExercise: "Impossibile aggiungere l'esercizio.",
+      updateDay: 'Impossibile aggiornare giornata: {message}',
+      deleteDay: 'Impossibile eliminare giornata: {message}',
+      missingDayExercise: 'Esercizio mancante nella giornata.',
+      updateDayExercise: "Impossibile aggiornare l'esercizio: {message}",
+      deleteDayExercise: "Impossibile eliminare l'esercizio: {message}",
+      loadMaxTests: 'Impossibile caricare i test massimali.',
+      loadMaxTestsWithMessage: 'Impossibile caricare i test massimali: {message}',
+    },
+    confirm: {
+      deleteExercise: 'Eliminare l’esercizio "{name}"?',
+      deletePlan: 'Eliminare questo piano?',
+      deleteDay: 'Eliminare questa giornata e i suoi esercizi?',
+      deleteDayExercise: "Rimuovere l'esercizio dalla giornata?",
+    },
+    dayCodes: {
+      MON: 'Lun',
+      TUE: 'Mar',
+      WED: 'Mer',
+      THU: 'Gio',
+      FRI: 'Ven',
+      SAT: 'Sab',
+      SUN: 'Dom',
+    },
+    planStatuses: {
+      active: 'attivo',
+      upcoming: 'in arrivo',
+      draft: 'bozza',
+      archived: 'archiviato',
+    },
+  },
+};
